More CO surv stopgaps (#10507)
# About the pull request

adds co equip overrides to literally every single insert

# Explain why it's good for the game

when doing #10270 I didn't consider that many maps have both CLF and
other team inserts, which would permit a CO to roll for survivor for the
CLF, causing them to spawn as a normal survivor in a team insert like
the UPP. This is behavior is undesirable IMO. Prompted by a CO surv
rolling on trijent when the UPP insert spawned and an admin having to
fix it.

As an aside I'm also trying to change the backend of how insert roles
are assigned in a different future to be more like ERTs with prefs to
make setting stuff like this up more intuitive and prevent COs/synths
from taking RP slots like the CLs in lv, solaris, and shivas, and make
it so that people can opt out of roles they don't want in these team
inserts. Until that's done, though, COs shouldn't spawn as truckers in
the upp ship, probably. My bad!


# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog
:cl:
fix: civilian survivor COs should no longer show up in team inerts like
the UPP on Trijent
/:cl: