M1911 weapon performance buffs (#10473)
# About the pull request

Weapon buffs for the M1911, particularly the M1911C and M1911 SOCOM, to
not make them peashooters especially compared to the stock counterpart
and the M4A3, which by all means should be a weaker version of it.

The following guns have been modified:
- M1911 (all variants): Increased firerate.
- M1911 SOCOM: Increased damage.
- M1911C (CO): Increased damage and slightly increased accuracy.

# Explain why it's good for the game

Tweaks are just minor to make them on par with other non-standard
pistols. Should not (in theory) make any variant "broken". At the very
least it'll actually be able to contend with standard issue pistols
which it wasn't able to do so before.

# Changelog

:cl:
balance: Increased weapon performance on all variants of the M1911,
particularly the newly-added M1911C.
/:cl: