Prevents throwing people upwards (#10513)
# About the pull request

Fixes #10248 

Adds a check to throwing code, to see if the object being thrown is a
living person, and if so, whether that person is being thrown upwards.

# Explain why it's good for the game

While throwing someone up in a parabolic arc is pretty funny, it also
allows you to bypass the climbing delay by instantly throwing someone
onto the roof.

# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

https://youtu.be/uwCRirtL02c

</details>


# Changelog

:cl:
del: You can no longer throw people upstairs.
/:cl: