Updates the messages sent for the hunting ground and youngblood ERTs (#10638)
# About the pull request
Title

<!-- Remove this text and explain what the purpose of your PR is.

Updates the messages sent for the hunting ground and youngblood ERTs

Mention if you have tested your changes. If you changed a map, make sure
you used the mapmerge tool.
If this is an Issue Correction, you can type "Fixes Issue #169420" to
link the PR to the corresponding Issue number #169420.

Remember: something that is self-evident to you might not be to others.
Explain your rationale fully, even if you feel it goes without saying.
-->

# Explain why it's good for the game
Makes these two erts more distinct 
# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog
:cl: Joe Lampost
add: Updates the messages sent for the hunting ground and youngblood
ERTs
/:cl: