Modifies brig vendor slogan (#10613)
# About the pull request

Removes a brig vendor slogan

# Explain why it's good for the game

Is the USCM anti-capitalist? This message feels out-of-place to me, but
I'm happy to close this PR if there's a valid reason for this slogan to
be here.


# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog
:cl:
del: Removes anti-capitalist vendor slogan from USCM vendor
/:cl: