Allows the MS-11 Smart Refill Tank to be refilled in the Pressurized Chem Dispenser (#10294)
# About the pull request

Allows you to put the MS-11 Smart Refill Tank in the pressurized chem
dispenser

# Explain why it's good for the game

The pressurized chem dispenser was created so you didn't have to haul an
entire reagent tank out of storage to fill up your pouch. This just
allows the refill tank the same function.

# Testing Photographs and Procedure

# Changelog

:cl:
add: Adds the ability to fill the MS-11 Smart Refill Tank in the
pressurized chemical dispensers
/:cl:

---------

Co-authored-by: Detective-Google <48196179+Detective-Google@users.noreply.github.com>