Being pulled off the tank is less punishing
The whole idea of receiving a debuff when you're pulled off of the tank is to penalize drivers who use the tank to spearhead a push and allow xenos to take advantage of poorly positioned marines atop the tank. A 3 second weaken is a bit too punishing and basically guarantees a cap. Now, the marine will get a shorter stun and a hefty slowdown, but they'll be able to fight back and can be saved if other marines pay attention.