Civilian Research (#9815)
# About the pull request

After an in-game poll that ran for 1+ week(s) the majority vote is for
Research to return to being civilians on loan to the USCM.


# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog
:cl:
add: The USS Almayer Research Department are now civilians again. They
do not answer to the CL.
/:cl:

---------

Co-authored-by: forest2001 <forest2001@cm-ss13.com>