Adds Non-Binary As A Gender Option (#10491)
# About the pull request
Adds Non-Binary as a gender option in character creation, it can also
roll on randomly generated humans with a 5% (Subject to maint approval,
I guess. PvE was fine with that number.) chance.
This should probably work, but be on the lookout for weird stuff with
naming and pronouns in messages from objs and such.
(Sort of a port of
[#1167](https://github.com/cmss13-devs/cmss13-pve/pull/1167), does it
count if it never got merged and got closed so I could open this
instead?)
# Explain why it's good for the game
Character stuff and diversity good.
# Changelog

:cl:
add: Adds Non-Binary as gender option in character creation and random
generation
ui: Adds a body presentation button to the PredPicker:tm:
/:cl:

---------

Co-authored-by: harryob <55142896+harryob@users.noreply.github.com>