Autodoc - Import surgery selection based on last bodyscan (#10308)
# About the pull request

Adds a button and associated functionality to use the last stored
bodyscan to select needed surgeries.
Added a filter to the import to exclude surgeries that haven't been
researched yet or are not fully supported by the code (missing limbs).

# Explain why it's good for the game

The logic was already being called and stored to the medical record by
the bodyscanner so why not put it to use?
Makes autodoc easier to use for non-doctors with medical skill, though
still of limited utility without research upgrades.

# Testing Photographs and Procedure

<details>
<summary>Screenshots & Videos</summary>


![autodoc_import](https://github.com/user-attachments/assets/c5120ed3-23a1-41d4-a4bc-36b5b51acd6f)
Not pictured: autodoc in the chatlog properly rejecting a surgery (IB)
because it's only partially upgraded.

</details>

# Changelog

:cl:
add: Added button to Autodoc to import surgery list generated by
bodyscanner
/:cl:

---------

Co-authored-by: harry <me@harryob.live>