multiz motion detector (#10593)
# About the pull request

motion detector detects contacts on upper and lower z, displays them
with queen eye icon, queen eye is detected only on the same level but
some dif icon might be needed in future. might add it to intel detectors
too not sure, might be a bit confusing there

# Explain why it's good for the game

THEY ARE IN THE ROOF! motion detector is one of the things that needs
update for multiz so here it is


# Testing Photographs and Procedure
<details>
tested on local with two accounts logged in
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog
:cl:
add: motion detector detects pings on higer and lower z level, uses
arrows to show their location
/:cl:

---------

Co-authored-by: vincibrv <vojtechvincibr@seznam.cz>
Co-authored-by: harryob <55142896+harryob@users.noreply.github.com>