QOL: simplified balloon messages (#10678)
# About the pull request

<!-- Remove this text and explain what the purpose of your PR is.

Mention if you have tested your changes. If you changed a map, make sure
you used the mapmerge tool.
If this is an Issue Correction, you can type "Fixes Issue #169420" to
link the PR to the corresponding Issue number #169420.

Remember: something that is self-evident to you might not be to others.
Explain your rationale fully, even if you feel it goes without saying.
-->

These are intended to be SHORT and PRECISE messages about game mechanics
shown to you when you do stuff not in a UI. They do not need
punctuation, capital letters, long wording, unnecessary adjectives,
especially not when being shown to all viewers.

the bimex alert was particularly egregious for flashing a three tile
wide message onto your screen when it was already a visible_message

# Explain why it's good for the game
reduces balloon message spam and blocking

# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog
:cl:
qol: simplified, reduced the length of, and in some cases removed
unnecessary balloon alerts.
/:cl:

Co-authored-by: Stan_Albatross <swt91a@gmail.com>
Co-authored-by: harry <me@harryob.live>