ERT/Preset QOL Tweaks and Fixes (#10652)
# About the pull request

Added presets for Visiting Officers who are Lt Colonel and Colonel
respectively.
Also added a preset for an MP Honor Guard (Not to be confused with the
Whiskey Outpost Honor Guard), and created a pamphlet in the HG kit box
so Military Police can update their title and HUD when they've been
designated Honor Guards.

# Explain why it's good for the game

A little bit of QOL. Makes it easier to understand why an MP is on the
frontline.


# Testing Photographs and Procedure



# Changelog
:cl:
add: Changed some ERT messages to be more related to what they are.
add: Added MP Honor Guard HUD and title pamphlet.
add: Added Visiting Lt Colonel and Colonel presets.
del: Removed an unused proc.
fix: Fixed an error with ARES WY IFF
/:cl:

---------

Co-authored-by: forest2001 <forest2001@cm-ss13.com>