Remove honor guard mp from role list (#10716)
# About the pull request

https://cdn.discordapp.com/attachments/604397850675380234/1420790458648498256/image.png?ex=68d6adba&is=68d55c3a&hm=7c853b3fda569367baeee33db69529f2ea529c2d105a10ae5c9f3e1fd01c4f99&
Bug
<!-- Remove this text and explain what the purpose of your PR is.

Mention if you have tested your changes. If you changed a map, make sure
you used the mapmerge tool.
If this is an Issue Correction, you can type "Fixes Issue #169420" to
link the PR to the corresponding Issue number #169420.

Remember: something that is self-evident to you might not be to others.
Explain your rationale fully, even if you feel it goes without saying.
-->

# Explain why it's good for the game
Fixes a bug
# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog
:cl:
fix: MP honor guard no longer shows up on the occupations preferences
list
/:cl: