Adds IFF to P9 SHARP (#10701)
# About the pull request

Adds IFF to the P9 SHARP

# Explain why it's good for the game

SHARP is currently considered a very underpowered spec, and the lack of
IFF prevents it from engaging in the first place; you are forced to
constantly cycle back and forth while you try to get targets for your
(rather slow) gun while marines try their best to run in front of you so
they can get an embedded metal rod in them to break their bones. Adding
IFF to the SHARP brings it at least in that regard on par with the other
fire support specs (like the GL or sniper) and lets it fire through the
crowd of PFCs like exotic pieces of equipment such as the UGL that's
attached to every MK2 by default.

# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog

:cl: 
balance: added IFF to P9 SHARP
/:cl: