M1911C Tweaks 2 (#10814)
# About the pull request

- Allows you to fire the M1911C into the air like all other CO guns.
- Allows you to purchase M1911C ammo from the CO vendor.
- Gives the M1911C impact rounds a `1.0 second` knockdown instead of a
`0.2 second` knockdown.

# Explain why it's good for the game

For a gun that only has one extra bullet and some more fire rate, you
cannot reliably BE with it at all. In terms of worries about juggling
xenos/humans with it's longer stun, you could do the same thing with the
Deagle if you really wanted to. I think its actually **easier** to miss
the juggle window with the higher firerate and lower stun time.

Firing into the air and the vendor are likely oversights, which are now
fixed.


# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog
:cl:
add: added M1911C ammo to the CO vendor.
balance: gave the M1911C a 1 second knockdown, up from 0.2 seconds.
fix: added the "fire in the air" ability to the M1911C to match all
other CO guns.
/:cl: