Changes the Commander headset to a senior command one (#10832)
# About the pull request
Changes the headset in the locker for the Commander of the operation to
an actual senior command headset.
# Explain why it's good for the game
The CO, XO and ASO already spawn with this headset. Anyone else that
would ever open the locker would be a game assigned Commander on lowpop
like a SO, PO or other lower-ranking marine. Considering the Commander
assigned by the game is responsible for the entire ship, including
Marine Law/MP related matters, it makes sense to ensure they have access
to all communications channels. When someone higher ranked than them
wakes up they will then be expected IC'ly to pass the senior command
headset along to them and go back to their original one.
# Changelog
:cl:
balance: Exchanged the CIC headset for the operations Commander for a
senior command headset
/:cl: