Fixes an overwatch console bino-blocker bypass (#10461)
# About the pull request

Fixes #9844

Missing a bino-check when "resuming" an overwatch, i.e. a camera has
been selected, but we're not currently set as an interactee on the
console. When pressing the button we're now set as an interactee, which
re-views the camera, bypassing the check.

# Explain why it's good for the game
# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog
:cl:
fix: Fixed a way to bypass the binocular blocker in the overwatch
console
/:cl: