Chain Of Command Additions (#10799)
# About the pull request

Adds All Officers and QM into the check for the Chain Of Command.
<!-- Remove this text and explain what the purpose of your PR is.

Mention if you have tested your changes. If you changed a map, make sure
you used the mapmerge tool.
If this is an Issue Correction, you can type "Fixes Issue #169420" to
link the PR to the corresponding Issue number #169420.

Remember: something that is self-evident to you might not be to others.
Explain your rationale fully, even if you feel it goes without saying.
-->

# Explain why it's good for the game
Clarity. There should be no reason why no one should know who is in
command of the ship, it wouldn't make sense.
# Testing Photographs and Procedure

<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog

:cl:
code: Changed the defines of Chain of Command
code: Added a check for doctors in to the list for Chain Of Command
announcement
fix: Fixes the Intel Officers to use the proper way of doing the check
and not checking via squad
/:cl: