Chances Claim - Dropship Loot Alteration + Minor Resin Change (#10861)
# About the pull request

Redistributes the loot within the west Dropship and evens it out with
the East dropship.

Removes the requirement to access the west dropship to open the east
dropship. You will need to C4 in to both dropships now.

Adds a few extra resin nodes near the reactor core entrance, and places
extra sticky resin by the east reactor entrance to slow down early game
intruders.

# Explain why it's good for the game

The loot redistribution is to shake up the loot meta and to de-emphasize
the concept that the westdropship is a prime loot location.

The extra weed and sticky resin is to provide early game xenos a small
defensive boon from survivor intruders, mainly the scout.


# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog
:cl:
maptweak: The loot inside the west dropship on Chances Claim has been
partially distributed to the east dropship. Furthermore, the east
dropship is now accessible via C4 instead of needing to be opened from
the west dropship.
maptweak: Additional resin nodes and sticky resin tiles have been placed
near the reactor core entrance. To slightly deter early game intruders.
/:cl:

Co-authored-by: Steelpoint <alexander.henley@hotmai.com>