Restore autowiki workflow back to node (#10855)
# About the pull request

This PR is a follow up to #10356 simply restoring a `bun` call back to
`node`. Possible the `npm install` could be changed to maybe `bun
install` but haven't heard back yet from Kash if this matters its node
or not (TG still has it using node)

# Testing Photographs and Procedure
You'll need to merge it and run autowiki to see.

# Changelog

No player facing changes.