If it looks like a fridge, acts like a fridge, and hums like a fridge... (#10827)
# About the pull request

Then it's a fridge.


# Explain why it's good for the game

It's very confusing to work out what is and isn't a 'fridge' according
to the nuke. They all look the same and only one counts.


# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog
:cl:
fix: The various types of freezer/fridges all now count for the nuke
protection.
/:cl:

Co-authored-by: forest2001 <forest2001@cm-ss13.com>