Reloading M240 flamethrower tanks from each other (#10812)
# About the pull request
Adds the ability to refill half-empty flamethrower tanks from one
another, for example, a tank with napalm from a tank with napalm or a
tank with B-gel from a tank with B-gel.
Fuel cannot be mixed.
<!-- Remove this text and explain what the purpose of your PR is.

Mention if you have tested your changes. If you changed a map, make sure
you used the mapmerge tool.
If this is an Issue Correction, you can type "Fixes Issue #169420" to
link the PR to the corresponding Issue number #169420.

Remember: something that is self-evident to you might not be to others.
Explain your rationale fully, even if you feel it goes without saying.
-->

# Explain why it's good for the game
QoL feature.
No more fuel tanks with 11 units of fuel scattered around the FOB.
# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.


https://github.com/user-attachments/assets/9ec438dc-6328-4844-8e82-ea18228ce39b


</details>


# Changelog
:cl:
qol: reloading M240 flamethrower tanks one by one

/:cl: