Multiz indicator (#9894)
# About the pull request

Adds an alert that lights up if there's open space above you.

# Explain why it's good for the game

Multi-Z being planned to be as significant as it is, there needs to be
an indicator for people on the ground that they can be seen from above.

Additionally, right now having Look-up be relegated to only a (default
unbound!) keybind, means that the majority of the playerbase doesn't
even know you CAN look up.

This PR aims to address both issues by adding a dynamic indicator that
lights up when there's something to look up at, and can be clicked to do
so.


# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>


https://github.com/user-attachments/assets/7965dc3a-b35b-433e-a085-80342bd52480

</details>


# Changelog


:cl:
add: There's a new indicator that shows whether there's a Multi-Z space
above you to look up at.
/:cl:

- [ ] TODO: make the indicator hide when on a level with no Z
transitions

---------

Co-authored-by: Git-Nivrak <59925169+Git-Nivrak@users.noreply.github.com>