Pred Onmob Fixes & Gun Belt Back Icon Support (#10927)
# About the pull request

This PR adds a check for updating your sprites when handling a gun
holster that's in your back slot (fixing a bug with the onmobs of the
hunting bow quiver), and also updates pred gear sprites across the board
to actually use an onmob when placed in the suit slot, rather than being
invisible.

I understand that in some cases we would rather have pred weapons in the
suit slot use a different unique onmob sprite, but right now they're
invisible, and making new sprites is beyond the scope of this PR.

# Explain why it's good for the game

Further bug fixes all around, and more consistency with items not being
invisible when placed in the suit slot.

# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>



https://github.com/user-attachments/assets/3860c8dc-a927-46d7-a8b7-d1a9c70377d0


</details>


# Changelog
:cl:
fix: hunting bow quiver back onmobs update properly
fix: pred weapons now visually show up on you when stored in the suit
slot
/:cl: