PO vendor qol and additions (#10819)
# About the pull request

General qol for the PO and DCC vendors.

* Adds a subtype of the pilot helmet without pilot visor preinstalled,
so you can now pick your helmet and visor color easily
* Removes the Tactical suit as a point buy and just makes it a suit
option for PO/GP. Same with the flak jacket.
* Updates general options that where missing due to oversigths like some
webbing colors and some new webbings.

# Explain why it's good for the game

The Vendors for PO havent been properly updated in a while. They got
additions, when someone rememberd to.
First of their helmet in the vendor had the PO visors preinstalled, and
ALSO an option for a custome color one including the basic one. So you
first have to fish out a screwdrive from somewhere else in the hangar.
Thats just kinda dumb. So the vendors get a subtype without preinstalled
PO visor.

Besides that the Tactical PO suit has been a point buy for ages for..
some reason. Its kinda like the only pure fashion item thats a point
buy, and quite an expensive one at that. So non of that, the suit and
jacket are now proper options the PO/GP can chose from.

Lastly there where just TONS of things missing from the vendors or it
was partially only? Like, PO got the leg pouches but not DCC. DCC had
black and brown drop pouches but not POs. Tons of that. Gave em the
shotgun choke as point uy attachment option, and the whistel too. If
rifelman can buy it, officers should definitly too.
Also sorted the attachments to now be sorted by attachment type. Same
with the webbings sligthly. Stuff being sorted well is always nice.

I am a bit confused about how to label stuff in the changelog. Im not
sure what counts as qol and what not. Like, no clue under what sorting
stuff falls. So... lemme know if i need to change stuff in the
changelog.

# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog
:cl:NHC
add: replaced the pilot helmet in the PO and DCC vendor with a version
that has no pilot visor pre-installed.
add: adds binoculars and some other minor items to the pilot vendors.
add: adds the tactical pilot officer flight suit and its armor as a free
option to the PO uniform vendor.
qol: sorts PO and DCC clothing vendor subcategories.
/:cl: