Moves SDTimeAward out of staff ranks. (#10895)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

# About the pull request

Moves the SDTimeAward out of the staff file and gives it the capability
to function as a community award as originally intended. No longer
requiring pseudo-staff assignment.

# Explain why it's good for the game

Removes the tiny bit of extra information given to people, and
standardises the checks for if you can play a role or not.


# Testing Photographs and Procedure
<!-- Include any screenshots/videos/debugging steps of the modified code
functioning successfully, ideally including edge cases. -->

<!-- !! If you are modifying sprites, you **must** include one or more
in-game screenshots or videos of the new sprites. !! -->

<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog

<!-- If your PR modifies aspects of the game that can be concretely
observed by players or admins you should add a changelog. If your change
does NOT meet this description, remove this section. Be sure to properly
label your changes in the changelog. Please note that maintainers freely
reserve the right to remove and add tags should they deem it
appropriate. You can attempt to finagle the system all you want, but
it's best to shoot for clear communication right off the bat. -->
<!-- If you add a name after the ':cl', that name will be used in the
changelog. You must add your CKEY after the CL if your GitHub name
doesn't match. Maintainers freely reserve the right to remove and add
tags should they deem it appropriate. -->

:cl:
add: Moved the preference for using playtime perks to settings tab
instead of xenos tab.
code: Added backend functionality for SDTimeAward community award to
function without being in the staff list.
admin: Added a preference for staff to decide if they use their own
playtime for ranks or not.
/:cl:

<!-- Both :cl:'s are required for the changelog to work! -->

---------

Co-authored-by: forest2001 <forest2001@cm-ss13.com>