Kutjevo CLF smuggler insert fix (#10939)
Fixes the CLF smuggler insert which was broken here:
https://github.com/cmss13-devs/cmss13/pull/8963

Reverts it to the previous version.

# About the pull request

Fixes the CLF smuggler insert map by reverting it back to the correct
version, which is correctly sized.

# Explain why it's good for the game

Someone made a mistake of not checking the maps, so the insert is broken
now, and cuts off a big amount of the insert and leaves a big part of
the new section in, very broken and very weird, this will fix it.

# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog
:cl: Zenith

fix: fixed the clf kutjevo insert by reverting it back to the correct
version.

/:cl: