Ticket Panel (Adminhelps & Mentorhelps)
This is a ticket panel intended to improve the staff side management of tickets. It provides a GUI for all admin and mentor tickets in one place, and allows you to see archived ones. There are quick access buttons for the player who is making the original request, and an easy system for labelling tickets with a subject for easy access.

Also expands mentorhelps so they have more function comparable to adminhelps.