Event Mob Spawning (#10403)
# About the pull request

This PR should make it significantly easier for staff to setup small to
mid scale events. It allows for people to join into event roles
similarly to Freed Mobs, with customisation as to the preset used.
Ghosts should spawn as their current character unless they've been in
the round already.

I will need my code checked and tested to ensure there's no way to
bypass it. Currently I'm not confident it is exploit proof.

# Explain why it's good for the game

Simplicity with events means more events, and hopefully more fun.

# Changelog
:cl:
add: Added event mob spawn landmarks, and a ghost verb to spawn as them.
add: You can only spawn as an event mob once per round, unless an admin
overrides this.
admin: Admins have freedom to setup these landmarks either in a map
insert or on-the-fly depending on which type is used. They're fairly
customisable and will allow for overrides of a few features on the
preset such as title and paygrade.
/:cl:

---------

Co-authored-by: forest2001 <forest2001@cm-ss13.com>
Co-authored-by: harry <me@harryob.live>