Minor Fix: Range Finder displays now displays Z level in the non TGUI mode. (#11112)
# About the pull request
As title: Adds Z info to the text output of the range finders which was
just an oversight.
<!-- Remove this text and explain what the purpose of your PR is.

Mention if you have tested your changes. If you changed a map, make sure
you used the mapmerge tool.
If this is an Issue Correction, you can type "Fixes Issue #169420" to
link the PR to the corresponding Issue number #169420.

Remember: something that is self-evident to you might not be to others.
Explain your rationale fully, even if you feel it goes without saying.
-->

# Explain why it's good for the game
Uhmmmm info good.
# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog
:cl: MistChristmas
fix: Rangefinder non-TGUI mode includes Z level info
/:cl: