CMLIVE!™️ (#10830)
the game automatically spectates people based on a priority system:

priority one: actively engaged in combat (causing damage) in the last 20
seconds
priority two: active players (talked or moved in the last 20 seconds) on
the ground map (pre hijack)
priority three: active players anywhere

when you are going to be spectated, you are given 10 seconds to cancel.
if you do, it’s given to someone else. this is not a permanent opt out

while being spectated, you have an extra "Handoff CMTV" verb if you want
rid of the uav

:cl:
server: added the ability to configure a ckey that will be automatically
moved to observe the most interesting players
server: added additional world/Topic endpoints to interface with CMTV
server: added the ability to link your Twitch account to the game
/:cl:

---------

Co-authored-by: harryob <55142896+harryob@users.noreply.github.com>
Co-authored-by: zzzmike <85382350+zzzmike@users.noreply.github.com>