Leadership Changes: Focus Mode Buffs (#10965)
# About the pull request

Part 1 of probably a trio of changes.

# Explain why it's good for the game

Currently, focus mode is very underwhelming, where at its current state,
only buffs accuracy, which in turn, only affects if the projectile can
successfully hit a mob or not

This PR buffs focus mode to include bonuses for both camera recoil and
scatter for the weapon itself, albeit not visible in the weapon stat
screen

You won't see much changes on lower leadership skill of course, but its
very noticeable on Leadership 3
# Testing Photographs and Procedure

It would be better to test out the buffs yourself, but I found Ol'
Painless to be friendlier for testing purposes

LEADERSHIP 3
https://youtu.be/oP2aYuzee_w

LEADERSHIP 2
https://youtu.be/P7dhvbNmm38

# Changelog
:cl:
balance: Focus mode now buffs both camera recoil and scatter alongside
the accuracy of a firearm.
/:cl: