Restore tacmap cooldown as a faction cooldown again (#11150)
# About the pull request

This PR is a follow up to #8963 which undid the global factional
cooldown for tacmaps. It also fixes a runtime that happens a lot.

# Explain why it's good for the game

Never was intended for a SO to run around CIC clicking every overwatch
computer to spam tacmap updates to turn a not live tacmap into a live
one.

# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

https://youtu.be/oHsC4bV6UJ4

</details>


# Changelog
:cl: Drathek
balance: Tacmap updates are a global cooldown per faction once again
/:cl: