Shield Code backend rewrite. (#9141)
# About the pull request

Shields are horridly complicated. Whew.

The intent of this PR is to rewrite the backend workings of all shield
blocks. That is specifically referring to hand-held shields for humans,
not xeno shields.

The reason for this is to make things more flexible, easier to see what
is doing what, and to allow for individual items to be more
appropriately fine-tuned.

# Explain why it's good for the game
# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog
:cl:
code: Changed how human hand-held blocking is handled.
code: Items that can be used as shields have a base percentage chance of
being used, reduced by factors such as if they need to be wielded (and
aren't) or not.
balance: Most entities that act as shields for humans are now
directional only, and cannot block things that are not in a roughly 75
degree arc of where the holder is facing. (Facing NORTH blocks from
NORTH-WEST, NORTH and NORTH-EAST).
balance: Machetes and other long blades have a 15% base chance of
blocking melee attacks, but cannot block ranged attacks.
balance: Xeno pounces can only be blocked by a small number of large
weapons.
/:cl:

---------

Co-authored-by: forest2001 <forest2001@cm-ss13.com>
Co-authored-by: Drathek <76988376+Drulikar@users.noreply.github.com>
Co-authored-by: fira <loyauflorian@gmail.com>